added the physical memory manager, started with paging

+++ src/kernel/include/pmm.h
@@ -7,8 +7,8 @@
 
 #define PMM_NO_MEM (void *)0xDEADC0DE
 
-void free_frame(uintptr_t addr, uint32_t frames);
-void* alloc_frame(uint32_t frames);
+void free_frame(uintptr_t addr, size_t frames);
+void* alloc_frame(size_t frames);
 void pmm_init(multiboot_info_t *mb_info);
